Refactor BSP module with dependency inversion and iterators
- Implement dependency inversion pattern using BspOps trait
- Create separate SerialBspOps and ParallelBspOps implementations
- Replace manual loops with iterator patterns throughout
- Maintain backward compatibility with existing Node API
- Add SplittingPlaneStrategy trait for algorithm flexibility
- Organize code into clean submodule structure:
  - traits.rs: Core trait definitions
  - node.rs: Node data structure only
  - serial.rs: Serial BSP operations
  - parallel.rs: Parallel BSP operations (with rayon)
  - mod.rs: Public API and backward compatibility
- Fix doctest import issues and create missing directories
- All tests passing (98 lib tests + 28 doc tests)

This refactoring separates algorithms from data structures,
enables easy testing of different BSP strategies, and follows
modern Rust patterns while maintaining full API compatibility.
