+++ src/mesh/bsp/node.rs
@@ -1,0 +1,41 @@
+//! BSP tree node data structure
+
+use crate::mesh::plane::Plane;
+use crate::mesh::polygon::Polygon;
+use std::fmt::Debug;
+
+/// A BSP tree node, containing polygons plus optional front/back subtrees
+#[derive(Debug, Clone)]
+pub struct Node<S: Clone> {
+    /// Splitting plane for this node *or* **None** for a leaf that
+    /// only stores polygons.
+    pub plane: Option<Plane>,
+
+    /// Polygons in *front* half‑spaces.
+    pub front: Option<Box<Node<S>>>,
+
+    /// Polygons in *back* half‑spaces.
+    pub back: Option<Box<Node<S>>>,
+
+    /// Polygons that lie *exactly* on `plane`
+    /// (after the node has been built).
+    pub polygons: Vec<Polygon<S>>,
+}
+
+impl<S: Clone + Send + Sync + Debug> Default for Node<S> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl<S: Clone + Send + Sync + Debug> Node<S> {
+    /// Create a new empty BSP node
+    pub const fn new() -> Self {
+        Self {
+            plane: None,
+            front: None,
+            back: None,
+            polygons: Vec::new(),
+        }
+    }
+}
